requirements.txt
flask
flask-restful
marshmallow
spacy
python-docx
openpyxl
